mqtt-gemhost-app/requirements.txt
paho-mqtt
secsgem